logs

+++ src/controller/Reference/Id-Entry.ts
@@ -15,6 +15,8 @@ import {
 import { IDGenerator, UpdateId } from "../../model/StoredProcedure";
 import { ExportToExcel } from "../../lib/exporttoexcel";
 import { mapDataBasedOnHeaders } from "../../lib/mapbaseonheader";
+import saveUserLogs from "../../lib/save_user_logs";
+import { saveUserLogsCode } from "../../lib/saveUserlogsCode";
 
 const ID_Entry = express.Router();
 
@@ -23,9 +25,13 @@ ID_Entry.post("/id-entry-client", async (req: Request, res: Response) => {
   const newMonth = ym.substring(0, 2);
   const newYear = ym.substring(2);
   req.body.createdAt = new Date();
+  delete req.body.mode;
+  delete req.body.search;
   try {
     await CreateClientEntry(req.body);
     await UpdateId("entry client", newCount, newMonth, newYear);
+    await saveUserLogs(req, req.body.entry_client_id, "add", "Entry Client");
+
     res.send({
       message: "Successfully Create New Client ID Entry",
       success: true,
@@ -42,9 +48,17 @@ ID_Entry.post("/id-entry-employee", async (req: Request, res: Response) => {
   const newYear = ym.substring(2);
 
   req.body.createdAt = new Date();
+  delete req.body.mode;
+  delete req.body.search;
   try {
     await CreateEmployeeEntry(req.body);
     await UpdateId("entry employee", newCount, newMonth, newYear);
+    await saveUserLogs(
+      req,
+      req.body.entry_employee_id,
+      "add",
+      "Entry Employee"
+    );
     res.send({
       message: "Successfully Create New Employee ID Entry",
       success: true,
@@ -60,9 +74,13 @@ ID_Entry.post("/id-entry-agent", async (req: Request, res: Response) => {
   const newMonth = ym.substring(0, 2);
   const newYear = ym.substring(2);
   req.body.createdAt = new Date();
+  delete req.body.mode;
+  delete req.body.search;
   try {
     await CreateAgentEntry(req.body);
     await UpdateId("entry agent", newCount, newMonth, newYear);
+    await saveUserLogs(req, req.body.entry_agent_id, "add", "Entry Agent");
+
     res.send({
       message: "Successfully Create New Agent ID Entry",
       success: true,
@@ -78,9 +96,18 @@ ID_Entry.post("/id-entry-fixed-assets", async (req: Request, res: Response) => {
   const newMonth = ym.substring(0, 2);
   const newYear = ym.substring(2);
   req.body.createdAt = new Date();
+  delete req.body.mode;
+  delete req.body.search;
   try {
     await CreateFixedAssetstEntry(req.body);
     await UpdateId("entry fixed assets", newCount, newMonth, newYear);
+    await saveUserLogs(
+      req,
+      req.body.entry_fixed_assets_id,
+      "add",
+      "Entry Fixed Assets"
+    );
+
     res.send({
       message: "Successfully Create New Fixed Assets ID Entry",
       success: true,
@@ -96,9 +123,18 @@ ID_Entry.post("/id-entry-supplier", async (req: Request, res: Response) => {
   const newMonth = ym.substring(0, 2);
   const newYear = ym.substring(2);
   req.body.createdAt = new Date();
+  delete req.body.mode;
+  delete req.body.search;
   try {
     await CreateSupplierEntry(req.body);
     await UpdateId("entry supplier", newCount, newMonth, newYear);
+    await saveUserLogs(
+      req,
+      req.body.entry_supplier_id,
+      "add",
+      "Entry Supplier"
+    );
+
     res.send({
       message: "Successfully Create New Supplier ID Entry",
       success: true,
@@ -114,9 +150,13 @@ ID_Entry.post("/id-entry-others", async (req: Request, res: Response) => {
   const newMonth = ym.substring(0, 2);
   const newYear = ym.substring(2);
   req.body.createdAt = new Date();
+  delete req.body.mode;
+  delete req.body.search;
   try {
     await CreateOtherEntry(req.body);
     await UpdateId("entry others", newCount, newMonth, newYear);
+    await saveUserLogs(req, req.body.entry_others_id, "add", "Entry Others");
+
     res.send({
       message: "Successfully Create New Other ID Entry",
       success: true,
@@ -129,6 +169,33 @@ ID_Entry.post("/id-entry-others", async (req: Request, res: Response) => {
 
 ID_Entry.post("/entry-update", async (req, res) => {
   try {
+    const list = [
+      { key: "entry_others_id", module: "Entry Others" },
+      { key: "entry_client_id", module: "Entry Client" },
+      { key: "entry_employee_id", module: "Entry Employee" },
+      { key: "entry_agent_id", module: "Entry Agent" },
+      { key: "entry_fixed_assets_id", module: "Entry Fixed Assets" },
+      { key: "entry_supplier_id", module: "Entry Supplier" },
+    ];
+    const obj = req.body;
+    const requester = { id: "", module: "" };
+    for (const key of Object.keys(obj)) {
+      if (list.map((itms) => itms.key).includes(key)) {
+        requester.id = obj[key];
+        requester.module = list.filter((itms) => itms.key === key)[0].module;
+        break;
+      }
+    }
+    if (
+      !(await saveUserLogsCode(req, "edit", requester.id, requester.module))
+    ) {
+      return res.send({ message: "Invalid User Code", success: false });
+    }
+
+    delete req.body.mode;
+    delete req.body.search;
+    delete req.body.userCodeConfirmation;
+
     await updateEntry(req.query.entry as string, req.body);
     res.send({ message: "Update Successfully", success: true });
   } catch (err: any) {
@@ -321,6 +388,17 @@ ID_Entry.get("/export-entry", async (req, res) => {
 
 ID_Entry.post("/entry-delete", async (req, res) => {
   try {
+    if (
+      !(await saveUserLogsCode(
+        req,
+        "delete ",
+        req.body.id,
+        `Entry ${req.query.entry}`
+      ))
+    ) {
+      return res.send({ message: "Invalid User Code", success: false });
+    }
+
     await deleteEntry(req.query.entry as string, req.body.id);
     res.send({
       success: true,
@@ -348,5 +426,9 @@ ID_Entry.get("/sub-account", async (req: Request, res: Response) => {
     res.send({ success: false, message: err.message });
   }
 });
+
+function capitalizeFirstLetter(string: string) {
+  return string.charAt(0).toUpperCase() + string.slice(1);
+}
 
 export default ID_Entry;
