+++ src/controller/Reference/policy-account.ts
@@ -8,12 +8,17 @@ import {
 } from "../../model/Reference/policy-account.model";
 import { mapDataBasedOnHeaders } from "../../lib/mapbaseonheader";
 import { ExportToExcel } from "../../lib/exporttoexcel";
+import { saveUserLogsCode } from "../../lib/saveUserlogsCode";
+import saveUserLogs from "../../lib/save_user_logs";
 
 const PolicyAccount = express.Router();
 
 PolicyAccount.post(
   "/add-policy-account",
   async (req: Request, res: Response) => {
+    delete req.body.mode;
+    delete req.body.search;
+    req.body.createdAt = new Date();
     if (await checkedAccountIsExisting(req.body.Account as string)) {
       return res.send({
         message: "This account is already exist",
@@ -21,12 +26,23 @@ PolicyAccount.post(
       });
     }
     await createPolicyAccount(req.body);
+    await saveUserLogs(req, req.body.Account, "add", "Policy Account");
     res.send({ message: "Create Policy Account Successfuly", success: true });
   }
 );
 PolicyAccount.post(
   "/update-policy-account",
   async (req: Request, res: Response) => {
+    if (
+      !(await saveUserLogsCode(req, "edit", req.body.Account, "Policy Account"))
+    ) {
+      return res.send({ message: "Invalid User Code", success: false });
+    }
+
+    delete req.body.mode;
+    delete req.body.search;
+    delete req.body.userCodeConfirmation;
+
     const findAccount = await checkedAccountIsExisting(
       req.body.Account as string
     );
@@ -36,7 +52,7 @@ PolicyAccount.post(
         success: false,
       });
     }
-    delete req.body.createdAt 
+    delete req.body.createdAt;
     updateValues(req.body);
     await updatePolicyAccount(req.body, findAccount.Account);
     res.send({ message: "Update Policy Account Successfuly", success: true });
@@ -45,6 +61,16 @@ PolicyAccount.post(
 PolicyAccount.post(
   "/delete-policy-account",
   async (req: Request, res: Response) => {
+    if (
+      !(await saveUserLogsCode(
+        req,
+        "delete",
+        req.body.Account,
+        "Policy Account"
+      ))
+    ) {
+      return res.send({ message: "Invalid Code", success: false });
+    }
     await deletePolicyAccount(req.body.Account as string);
     res.send({ message: "Delete Policy Account Successfuly", success: true });
   }
@@ -53,10 +79,10 @@ PolicyAccount.get(
   "/get-policy-account",
   async (req: Request, res: Response) => {
     const { policySearch } = req.query;
-    const policy:any = await searchPolicy(policySearch as string) 
-    policy.map((obj:any)=>{
-      return updateValues(obj)
-    })
+    const policy: any = await searchPolicy(policySearch as string);
+    policy.map((obj: any) => {
+      return updateValues(obj);
+    });
     res.send({
       message: "Get Policy Account Successfuly",
       success: true,
@@ -69,11 +95,11 @@ PolicyAccount.get(
   "/search-policy-account",
   async (req: Request, res: Response) => {
     const { policySearch } = req.query;
-    const policy:any = await searchPolicy(policySearch as string);
+    const policy: any = await searchPolicy(policySearch as string);
 
-    policy.map((obj:any)=>{
-      return updateValues(obj)
-    })
+    policy.map((obj: any) => {
+      return updateValues(obj);
+    });
     res.send({
       message: "Search Policy Account Successfuly",
       success: true,
@@ -97,18 +123,8 @@ function updateValues(obj: any) {
 PolicyAccount.get("/export-policy-account", async (req, res) => {
   const entryHeaders: any = {
     Policy: {
-      header: [
-        "Account",
-        "Account Code",
-        "Description",
-        "Created At"
-      ],
-      row: [
-        "Account",
-        "AccountCode",
-        "Description",
-        "createdAt"
-      ],
+      header: ["Account", "Account Code", "Description", "Created At"],
+      row: ["Account", "AccountCode", "Description", "createdAt"],
     },
   };
   const { policySearch, isAll } = req.query;
@@ -118,13 +134,13 @@ PolicyAccount.get("/export-policy-account", async (req, res) => {
     data = mapDataBasedOnHeaders(
       (await searchPolicy("", true)) as Array<any>,
       entryHeaders,
-      'Policy'
+      "Policy"
     );
   } else {
     data = mapDataBasedOnHeaders(
       (await searchPolicy(policySearch as string)) as Array<any>,
       entryHeaders,
-      'Policy'
+      "Policy"
     );
   }
